requirements.txt
ffmpy==0.3.0
i18nice==0.6.2
Pillow==9.2.0
pyheif==0.7.1
python-dotenv==0.21.1
python-telegram-bot==20.3
rlottie_python==1.1.1
